+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Make target labels more easily configurable in `values.yaml`.
+
 ## [0.1.0] - 2021-11-26
 
 ### Added